+++ frontend/svelte/src/lib/components/accounts/HardwareWalletConnect.svelte
@@ -7,7 +7,7 @@
   import {
     ADD_ACCOUNT_CONTEXT_KEY,
     type AddAccountContext,
-  } from "../../stores/add-account.store";
+  } from "../../types/add-account.context";
   import { createEventDispatcher, getContext } from "svelte";
   import type { LedgerIdentity } from "../../identities/ledger.identity";
   import { busy, startBusy, stopBusy } from "../../stores/busy.store";
