+++ frontend/svelte/src/tests/lib/components/accounts/NewTransactionTest.svelte
@@ -3,14 +3,14 @@
   import {
     NEW_TRANSACTION_CONTEXT_KEY,
     TransactionContext,
-    transactionStore,
-  } from "../../../../lib/stores/transaction.store";
+  } from "../../../../lib/types/transaction.context";
+  import { mockTransactionStore } from "../../../mocks/transaction.store.mock";
 
   export let testComponent: typeof SvelteComponent;
   export let nextCallback: () => void | undefined = undefined;
 
   setContext<TransactionContext>(NEW_TRANSACTION_CONTEXT_KEY, {
-    store: transactionStore,
+    store: mockTransactionStore,
     next: () => nextCallback?.(),
   });
 </script>
