+++ frontend/svelte/src/tests/mocks/add-account.store.mock.ts
@@ -1,0 +1,7 @@
+import { writable } from "svelte/store";
+import type { AddAccountStore } from "../../lib/types/add-account.context";
+
+export const addAccountStoreMock = writable<AddAccountStore>({
+  type: undefined,
+  hardwareWalletName: undefined,
+});
